users basic analytics #47 (fix bug)
