+++ backend/server/src/controllers/dashboard/dashboardController.js
@@ -1,124 +1,24 @@
 const { connectToDatabase } = require("../../configuration/dbConfig");
 const { ObjectId } = require("mongodb");
+const { calculateUserStats } = require("../../services/userStatsService");
 
-const calculateUserStats = async (db, userObjectId) => {
-  const [storyLikesResult, confessionLikesResult] = await Promise.all([
-    db
-      .collection("storyLikes")
-      .aggregate([
-        {
-          $lookup: {
-            from: "stories",
-            localField: "storyId",
-            foreignField: "_id",
-            as: "story",
-          },
-        },
-        {
-          $unwind: "$story",
-        },
-        {
-          $match: {
-            "story.authorId": userObjectId,
-            "story.deletedAt": null,
-          },
-        },
-        {
-          $count: "totalLikes",
-        },
-      ])
-      .toArray(),
-    db
-      .collection("confessionLikes")
-      .aggregate([
-        {
-          $lookup: {
-            from: "confessions",
-            localField: "confessionId",
-            foreignField: "_id",
-            as: "confession",
-          },
-        },
-        {
-          $unwind: "$confession",
-        },
-        {
-          $match: {
-            "confession.authorId": userObjectId,
-            "confession.deletedAt": null,
-          },
-        },
-        {
-          $count: "totalLikes",
-        },
-      ])
-      .toArray(),
-  ]);
+const getAuthenticatedUserObjectId = (req, res) => {
+  if (!ObjectId.isValid(req.user?.userId)) {
+    res.status(401).json({ message: "Invalid user id" });
+    return null;
+  }
 
-  const [storyWordsResult, confessionWordsResult] = await Promise.all([
-    db
-      .collection("stories")
-      .aggregate([
-        {
-          $match: {
-            authorId: userObjectId,
-            deletedAt: null,
-          },
-        },
-        {
-          $group: {
-            _id: null,
-            totalWords: { $sum: "$wordCount" },
-            storyCount: { $sum: 1 },
-          },
-        },
-      ])
-      .toArray(),
-    db
-      .collection("confessions")
-      .aggregate([
-        {
-          $match: {
-            authorId: userObjectId,
-            deletedAt: null,
-          },
-        },
-        {
-          $group: {
-            _id: null,
-            totalWords: { $sum: "$wordCount" },
-            confessionCount: { $sum: 1 },
-          },
-        },
-      ])
-      .toArray(),
-  ]);
-
-  const storyLikes = storyLikesResult[0]?.totalLikes || 0;
-  const confessionLikes = confessionLikesResult[0]?.totalLikes || 0;
-  const storyWords = storyWordsResult[0]?.totalWords || 0;
-  const confessionWords = confessionWordsResult[0]?.totalWords || 0;
-  const storyCount = storyWordsResult[0]?.storyCount || 0;
-  const confessionCount = confessionWordsResult[0]?.confessionCount || 0;
-
-  return {
-    totalLikes: storyLikes + confessionLikes,
-    totalWords: storyWords + confessionWords,
-    totalPosts: storyCount + confessionCount,
-    breakdown: {
-      storyLikes,
-      confessionLikes,
-      storyWords,
-      confessionWords,
-      storyCount,
-      confessionCount,
-    },
-  };
+  return new ObjectId(req.user.userId);
 };
 
 const getDashboardStats = async (req, res) => {
   try {
-    const userObjectId = new ObjectId(req.user.userId);
+    const userObjectId = getAuthenticatedUserObjectId(req, res);
+
+    if (!userObjectId) {
+      return;
+    }
+
     const db = await connectToDatabase();
     const stats = await calculateUserStats(db, userObjectId);
 
@@ -134,7 +34,12 @@ const getDashboardStats = async (req, res) => {
 
 const getDashboardStories = async (req, res) => {
   try {
-    const userObjectId = new ObjectId(req.user.userId);
+    const userObjectId = getAuthenticatedUserObjectId(req, res);
+
+    if (!userObjectId) {
+      return;
+    }
+
     const {
       status = "all",
       visibility = "all",
@@ -207,7 +112,12 @@ const getDashboardStories = async (req, res) => {
 
 const getDashboardConfessions = async (req, res) => {
   try {
-    const userObjectId = new ObjectId(req.user.userId);
+    const userObjectId = getAuthenticatedUserObjectId(req, res);
+
+    if (!userObjectId) {
+      return;
+    }
+
     const { sortBy = "date", order = "desc", page = 1, limit = 20 } = req.query;
 
     const parsedPage = Number.parseInt(page, 10) || 1;
